+++ lib/stellar_client/requests/send_payment_request.rb
@@ -1,4 +1,4 @@
-module BridgeClient
+module StellarClient
   class SendPaymentRequest < BaseRequest
 
     BODY_ATTRS = %i[