+++ spec/spec_helper.rb
@@ -1,17 +1,17 @@
 $LOAD_PATH.unshift File.expand_path("../../lib", __FILE__)
-require "virtus-matchers"
+require "virtus/matchers/rspec"
 require "pry-byebug"
 require "pathname"
 require "active_support/core_ext/hash/indifferent_access"
 require "active_support/json"
 require 'vcr'
-require "bridge_client"
+require "stellar_client"
 
 SPEC_DIR = Pathname.new(File.dirname(__FILE__))
 CONFIG_YML = SPEC_DIR.join("config.yml")
 CONFIG = YAML.load_file(CONFIG_YML).with_indifferent_access
 
-BridgeClient.configure do |c|
+StellarClient.configure do |c|
   c.host = CONFIG[:host]
 end