+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Renamed from BridgeClient to StellarClient
+
 ## [0.4.0] - 2017-04-05
 ### Added
 - Expose response's `code` (as `code_str`) and `message`